refactor: breakdown NavbarHelper to Navigable modules
