+++ app/helpers/application_helper.rb
@@ -2,14 +2,7 @@
 
 module ApplicationHelper
   include StyleHelper
-
-  def anonymous_dom_id(seed_string)
-    "#{slugify_for_dom(seed_string)}--#{SecureRandom.hex(4)}"
-  end
-
-  def slugify_for_dom(name)
-    name.to_s.parameterize(separator: '-')
-  end
+  include DynamicDomElementCapable
 
   def modal_dom_id(resource, content_type: nil)
     raise ArgumentError, "#{resource.class.name} MUST be Renderable" \
